refactor into jdk_compatibility_tests and perftests
