+++ integration_tests/jdk_compatibility_tests/src/test/java/io/fury/integration_tests/JDKCompatibilityTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.fury.integration_tests.state;
+package io.fury.integration_tests;
 
 import io.fury.Fury;
 import io.fury.Language;
