+++ integration_tests/perftests/src/main/proto/bench.proto
